data/gameplay_optimization/tools/pyproject.toml
[project]
name = "opt-gameplay"
version = "0.1.0"
description = "LLM-driven gameplay optimization tools for Arc Citadel"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "dspy-ai>=2.5.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
opt-gameplay = "opt_gameplay.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opt_gameplay"]
